Home

SimpleMock

SimpleMock.SimpleMock — Module.

A basic mocking module, inspired by Python's unittest.mock and implemented with Cassette.

source

The Mock object

SimpleMock.Mock — Type.

Mock(; return_value=Mock(), side_effect=nothing)

Create a new mocking object that can act as a replacement for a function.

Return Value

Use the return_value keyword to set the value to be returned upon calling the mock. By default, the return value is a new Mock.

Side Effects

Use the side_effect keyword to set a side effect to occur upon calling the mock.

If the value is an Exception, then the exception is thrown.

If the value is callable, then it is called with the same arguments and keywords.

If the value is a Vector, then each call uses the next element.

Any other value except nothing is returned without modification.

source

SimpleMock.calls — Function.

calls(::Mock) -> Vector{Call}

Return the call history of the Mock.

source

SimpleMock.ncalls — Function.

ncalls(::Mock) -> Int

Return the number of times that the Mock has been called.

source

SimpleMock.called — Function.

called(::Mock) -> Bool

Return whether or not the Mock has been called.

source

SimpleMock.called_once — Function.

called_once(::Mock) -> Bool

Return whether or not the Mock has been called exactly once.

source

SimpleMock.called_with — Function.

called_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock has been called with the given arguments.

source

SimpleMock.called_once_with — Function.

called_once_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock has been called exactly once with the given arguments.

source

SimpleMock.Call — Type.

Call(args...; kwargs...)

Represents a function call.

source

SimpleMock.has_call — Function.

has_call(::Mock, ::Call) -> Bool

Similiar to called_with, but using a Call.

source

SimpleMock.has_calls — Function.

has_calls(::Mock, ::Calls...) -> Bool

Return whether or not the Mock has a particular ordered sequence of Calls.

source

SimpleMock.reset! — Function.

reset!(::Mock)

Reset a Mock's call history. Side effects and return values are preserved.

source

The mock Function

SimpleMock.mock — Function.

mock(f::Function, args...; filters::Vector{<:Function}=Function[])

Run f with specified functions mocked out.

Note

Keyword arguments to mocked functions are not supported. If you call a mocked function with keyword arguments, it will dispatch to the original function. For more details, see Cassette#48.

Examples

Mocking a single function:

f(x) = x + 1
mock(+) do plus
    @assert plus isa Mock
    @assert f(0) != 1  # The call to + is mocked.
    @assert called_once_with(p, 0, 1)
end

Mocking a function with a custom Mock:

mock((+) => Mock(; return_value=1)) do plus
    @assert 1 + 1 == 1
    @assert called_once_with(plus, 1, 1)
end

Mocking methods that match a given signature:

mock((+, Float64, Float64) => Mock(; side_effect=(a, b) -> 2a + 2b)) do plus
    @assert 1 + 1 == 2
    @assert 2.0 + 2.0 == 8
    @assert called_once_with(plus, 2.0, 2.0)
end

Mocking with something other than a Mock:

mock((+) => (a, b) -> 2a + 2b) do _plus
    @assert 1 + 2 == 6
end

Using Filters

Oftentimes, you mock a function with a very specific idea of where you want that mocking to happen. It can be confusing when a call you didn't anticipate gets mocked somewhere deep in the call stack, botching everything. To avoid this, you can use filter functions like so:

f(x, y) = x + y
g(x, y) = f(x, y)
mock((+) => Mock(; side_effect=(a, b) -> 2a + 2b); filters=[max_depth(2)]) do plus
    @assert f(1, 2) == 6  # The call depth of + here is 2.
    @assert g(3, 4) == 7  # Here, it's 3.
    @assert called_once_with(plus, 1, 2)
end

Filter functions take a single argument of type Metadata. If any filter rejects, then mocking is not performed. See Filter Functions for a list of included filters, as well as building blocks for you to create your own.

source

Filter Functions

SimpleMock.max_depth — Function.

max_depth(n::Int) -> Function

Create a filter that rejects when the current call depth is greater than n.

source

SimpleMock.min_depth — Function.

min_depth(n::Int) -> Function

Create a filter that rejects when the current call depth is less than n.

source

SimpleMock.excluding — Function.

excluding(args...) -> Function

Create a filter that rejects when the calling function or module is in args.

source

SimpleMock.including — Function.

including(args...) -> Function

Create a filter that rejects when the calling function or module is not in args.

source

SimpleMock.Metadata — Type.

Container for mocks and bookkeeping data. Instances of this type are aware of the call depth (current_depth) and the current function/module (current_function/current_module).

All filter functions take a single argument of this type.

source

SimpleMock.current_depth — Function.

current_depth(::Metadata) -> Int

Return the current call depth (the size of the call stack). The depth is always positive, so the first function entered has a depth of 1.

source

SimpleMock.current_function — Function.

current_function(::Metadata) -> Any

Return the current function (or other callable thing). In this case, "current" refers, somewhat counterintuitively, not to the function about to be called, but to the function that is about to call it.

To illustrate:

f(x) = x
g(x) = f(x)
x = g(1)

In this case, when the call to f is reached, the function that is calling f is g. Therefore, the "current" function is g.

source

SimpleMock.current_module — Function.

current_module(m::Metadata) -> Module

Return the current module, where "current" has the same definition as in current_function.

source

Index

SimpleMock.SimpleMock

SimpleMock.Call

SimpleMock.Metadata

SimpleMock.Mock

SimpleMock.called

SimpleMock.called_once

SimpleMock.called_once_with

SimpleMock.called_with

SimpleMock.calls

SimpleMock.current_depth

SimpleMock.current_function

SimpleMock.current_module

SimpleMock.excluding

SimpleMock.has_call

SimpleMock.has_calls

SimpleMock.including

SimpleMock.max_depth

SimpleMock.min_depth

SimpleMock.mock

SimpleMock.ncalls

SimpleMock.reset!