Home

SimpleMock

SimpleMock.SimpleMock — Module.

A basic mocking module, inspired by Python's unittest.mock and implemented with Cassette.

Usage

For usage examples, see mock.

Differences from unittest.mock

SimpleMock only implements mocking of function calls, as opposed to arbitrary monkey-patching.

Neither getfield nor setfield! is not implemented for the default Mock object.

source

SimpleMock.mock — Function.

mock(f::Function[, ctx::Symbol], args...)

Run f with specified functions replaced with Mocks.

Warning

There are a few issues with this function, see them on GitHub.

Examples

Mocking a single function:

mock(get) do get
    Base.get(1)  # Would normally throw a `MethodError`.
    @assert called_once_with(get, 1)
end

Mocking a function with a custom Mock:

mock(get => Mock(; return_value=1)) do get
    @assert Base.get(1) == 1
    @assert called_once_with(get, 1)
end

Reusing A Context

TODO: doc about reusing Context types.

source

SimpleMock.Mock — Type.

Mock(; return_value=Mock(), side_effect=nothing)

Create a new mocking object, which behaves similarly to Python's Mock.

Return Value

Use the return_value keyword to set the value to be returned upon calling the mock. By default, the return value is a new Mock.

Side Effects

Use the side_effect keyword to set a side effect to occur upon calling the mock.

If the value is an Exception, then the exception is thrown.

If the value is a function, then it is called with the same arguments and keywords.

If the value is a Vector, then each call uses the next element.

Any other value is returned without modification.

source

SimpleMock.calls — Function.

calls(::Mock) -> Vector{Call}

Return the call history of the Mock.

source

SimpleMock.ncalls — Function.

ncalls(::Mock) -> Int

Return the number of times that the Mock has been called.

source

SimpleMock.called — Function.

called(::Mock) -> Bool

Return whether or not the Mock has been called.

source

SimpleMock.called_once — Function.

called_once(::Mock) -> Bool

Return whether or not the Mock has been called exactly once.

source

SimpleMock.called_with — Function.

called_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock has been called with the given arguments.

source

SimpleMock.called_once_with — Function.

called_once_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock has been called exactly once with the given arguments.

source

SimpleMock.Call — Type.

Call(args...; kwargs...)

Represents a function call.

source

SimpleMock.has_call — Function.

has_call(::Mock, ::Call) -> Bool

Similiar to called_with, but using a Call.

source

SimpleMock.has_calls — Function.

has_calls(::Mock, ::Calls...) -> Bool

Return whether or not the Mock has a particular ordered sequence of Calls.

source

SimpleMock.reset! — Function.

reset!(::Mock)

Reset a Mock's call history and internal variables. Side effects and return values are preserved.

source

SimpleMock.SimpleMock

SimpleMock.Call

SimpleMock.Mock

SimpleMock.called

SimpleMock.called_once

SimpleMock.called_once_with

SimpleMock.called_with

SimpleMock.calls

SimpleMock.has_call

SimpleMock.has_calls

SimpleMock.mock

SimpleMock.ncalls

SimpleMock.reset!