Home

SimpleMock

SimpleMock.SimpleMock — Module.

A basic mocking module, inspired by Python's unittest.mock and implemented with Cassette.

source

The Mock object

SimpleMock.Mock — Type.

Mock(; return_value=Mock(), side_effect=nothing)

Create a new mocking object that can act as a replacement for a function.

Return Value

Use the return_value keyword to set the value to be returned upon calling the mock. By default, the return value is a new Mock.

Side Effects

Use the side_effect keyword to set a side effect to occur upon calling the mock.

If the value is an Exception, then the exception is thrown.

If the value is a function, then it is called with the same arguments and keywords.

If the value is a Vector, then each call uses the next element.

Any other value except nothing is returned without modification.

source

SimpleMock.calls — Function.

calls(::Mock) -> Vector{Call}

Return the call history of the Mock.

source

SimpleMock.ncalls — Function.

ncalls(::Mock) -> Int

Return the number of times that the Mock has been called.

source

SimpleMock.called — Function.

called(::Mock) -> Bool

Return whether or not the Mock has been called.

source

SimpleMock.called_once — Function.

called_once(::Mock) -> Bool

Return whether or not the Mock has been called exactly once.

source

SimpleMock.called_with — Function.

called_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock has been called with the given arguments.

source

SimpleMock.called_once_with — Function.

called_once_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock has been called exactly once with the given arguments.

source

SimpleMock.Call — Type.

Call(args...; kwargs...)

Represents a function call.

source

SimpleMock.has_call — Function.

has_call(::Mock, ::Call) -> Bool

Similiar to called_with, but using a Call.

source

SimpleMock.has_calls — Function.

has_calls(::Mock, ::Calls...) -> Bool

Return whether or not the Mock has a particular ordered sequence of Calls.

source

SimpleMock.reset! — Function.

reset!(::Mock)

Reset a Mock's call history. Side effects and return values are preserved.

source

The mock Function

SimpleMock.mock — Function.

mock(f::Function[, ctx::Symbol], args...; filters::Vector{<:Function}=Function[])

Run f with specified functions mocked out.

Examples

Mocking a single function:

mock(print) do p
    @assert p isa Mock
    println("!")  # This won't output anything.
    @assert called_once_with(p, stdout, "!", '\n')
end

Mocking a function with a custom Mock:

mock((+) => Mock(; return_value=1)) do plus
    @assert 1 + 1 == 1
    @assert called_once_with(plus, 1, 1)
end

Mocking methods that match a given signature:

mock((+, Float64, Float64) => Mock(; side_effect=(a, b) -> 2a + 2b)) do plus
    @assert 1 + 1 == 2
    @assert 2.0 + 2.0 == 8
    @assert called_once_with(plus, 2.0, 2.0)
end

Mocking with something other than a Mock:

mock((+) => (a, b) -> 2a + 2b) do _plus
    @assert 1 + 2 == 6
end

Using Filters

Oftentimes, you mock a function with a very specific idea of where you want that mocking to happen. It can be confusing when a call you didn't anticipate gets mocked somewhere deep in the call stack, botching everything. To avoid this, you can use filter functions, like so:

f(x) = print(x)
g(x) = f(x)
mock(print; filters=[max_depth(2)]) do p
    f("this won't print")  # The call depth of print here is 2.
    g("this will print")   # Here, it's 3.
    @assert called_once_with(p, "this won't print")
end

Filter functions take a single argument of type Metadata. See Filter Functions for a list of included filters, as well as building blocks for you to create your own.

Performance Tips

Avoid Printing

Printing is, for whatever reason, glacially slow inside of mock blocks. To illustrate:

julia> @time mock(println, log)
Mock{Symbol,Nothing}(Symbol("##390"), Call[], Symbol("##371"), nothing)
  7.389156 seconds (19.52 M allocations: 1.029 GiB, 7.71% gc time)

julia> @time println(mock(identity, log))
Mock{Symbol,Nothing}(Symbol("##394"), Call[], Symbol("##371"), nothing)
  0.136950 seconds (120.96 k allocations: 6.507 MiB

Unfortunately, this includes the display of failed @tests, so it's wise to avoid making assertions in the mocked environment.

#= bad:  =# mock(lg -> @test(!called(lg), log)
#= good: =# @test !called(mock(identity, log))

The second strategy is orders of magnitude faster than the first when the test fails, and it's also faster when the test passes.

Don't Filter Unless Necessary

Filtering introduces significant bookkeeping overhead. Avoid it whenever possible!

Reuse Your Contexts

Under the hood, this function creates a new Cassette Context on every call by default. This provides a nice clean mocking environment, but it can be slow to create and call new types and methods over and over. If you find yourself repeatedly mocking the same set of functions, you can specify a context name to reuse that context like so:

julia> ctx = gensym();

# The first time takes a little while.
julia> @time mock(g -> @assert(!called(g)), ctx, get)
  0.156221 seconds (171.93 k allocations: 9.356 MiB)

# But the next time is faster!
julia> @time mock(g -> @assert(!called(g)), ctx, get)
  0.052324 seconds (27.38 k allocations: 1.437 MiB)

Be careful though! If you call a function that you've previously mocked but are not currently mocking, you'll run into trouble:

julia> f(s) = strip(uppercase(s));
julia> ctx = gensym();

julia> mock(_g -> f(" hi "), ctx, strip => s -> "hi");
julia> mock(_g -> f(" hi "), ctx, uppercase => s -> " HI ")
ERROR: KeyError: key (strip, Vararg{Any,N} where N) not found

source

Filter Functions

SimpleMock.max_depth — Function.

min_depth(n::Int) -> Function

Create a filter that rejects when the current call depth is greater than n.

source

SimpleMock.min_depth — Function.

min_depth(n::Int) -> Function

Create a filter that rejects when the current call depth is less than n.

source

SimpleMock.excluding — Function.

excluding(args...) -> Function

Create a filter that rejects when the calling function or module is in args.

source

SimpleMock.including — Function.

including(args...) -> Function

Create a filter that rejects when the calling function or module is not in args.

source

SimpleMock.Metadata — Type.

Container for mocks and bookkeeping data. Instances of this type are aware of the call depth (current_depth) and the current function/module (current_function/current_module).

All filter functions take a single argument of this type.

source

SimpleMock.current_depth — Function.

current_depth(::Metadata) -> Int

Return the current call depth (the size of the call stack). The depth is always positive, so the first function entered has a depth of 1.

source

SimpleMock.current_function — Function.

current_function(::Metadata) -> Any

Return the current function (or other callable thing). In this case, "current" refers, somewhat counterintuitively, not to the function about to be called, but to the function that is about to call it.

To illustrate:

f(x) = x
g(x) = f(x)
x = g(1)

In this case, when the call to f is reached, the function that is calling f is g. Therefore, the "current" function is g.

source

SimpleMock.current_module — Function.

current_module(m::Metadata) -> Module

Return the current module, where "current" has the same definition as in current_function.

source

SimpleMock.SimpleMock

SimpleMock.Call

SimpleMock.Metadata

SimpleMock.Mock

SimpleMock.called

SimpleMock.called_once

SimpleMock.called_once_with

SimpleMock.called_with

SimpleMock.calls

SimpleMock.current_depth

SimpleMock.current_function

SimpleMock.current_module

SimpleMock.excluding

SimpleMock.has_call

SimpleMock.has_calls

SimpleMock.including

SimpleMock.max_depth

SimpleMock.min_depth

SimpleMock.mock

SimpleMock.ncalls

SimpleMock.reset!