SimpleMock

SimpleMock.SimpleMock — Module

A basic mocking module, inspired by Python's unittest.mock and implemented with Cassette.

The Mock object

SimpleMock.Mock — Type

Mock([effect])

Create a new mocking object that can act as a replacement for a function.

Effects

Use the effect argument to determine what happens upon calling the mock.

If the value is callable, then it is called with the same arguments and keywords.

If the value is an Exception, then the exception is thrown.

If the value is a Vector, then each call consumes the next element. Nested Vectors are returned, but callables and exceptions are treated as explained above.

Any other value is returned without modification.

By default, calling a Mock returns a new Mock.

SimpleMock.calls — Function

calls(::Mock) -> Vector{Call}

Return the call history of the Mock.

SimpleMock.last_call — Function

last_call(::Mock) -> Call

Return the last call of the Mock.

SimpleMock.ncalls — Function

ncalls(::Mock) -> Int

Return the number of times that the Mock has been called.

SimpleMock.called — Function

called(::Mock) -> Bool

Return whether or not the Mock has been called.

SimpleMock.called_once — Function

called_once(::Mock) -> Bool

Return whether or not the Mock has been called exactly once.

SimpleMock.called_with — Function

called_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock has been called with the given arguments.

SimpleMock.called_last_with — Function

called_last_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock was last called with the given arguments.

SimpleMock.called_once_with — Function

called_once_with(::Mock, args...; kwargs...) -> Bool

Return whether or not the Mock has been called exactly once with the given arguments.

SimpleMock.Call — Type

Call(args...; kwargs...)

Represents a function call.

SimpleMock.has_call — Function

has_call(::Mock, ::Call) -> Bool

Similiar to called_with, but using a Call.

SimpleMock.has_calls — Function

has_calls(::Mock, ::Call...) -> Bool

Return whether or not the Mock has a particular ordered sequence of Calls.

SimpleMock.reset! — Function

reset!(::Mock)

Reset a Mock's call history. Its effect is preserved.

SimpleMock.Predicate — Type

Predicate(f)

A predicate that can be used instead of literal call arguments, useful for situations where the exact value of arguments is not known. The function f is called with the observed argument, and must return a Bool.

Example

m = Mock()
m(rand(1:10))
@assert called_with(m, Predicate(in(1:10)))

The mock Function

SimpleMock.mock — Function

mock(f::Function[, ctx::Symbol], args...; filters::Vector{<:Function}=Function[])

Run f with specified functions mocked out.

Note

Mocking functions with keyword arguments is only partially supported. See the "Keyword Arguments" section below for more details.

Examples

Mocking a single function:

f(x) = x + 1
mock(+) do plus
    @assert plus isa Mock
    @assert f(0) != 1  # The call to + is mocked.
    @assert called_once_with(plus, 0, 1)
end

Mocking a function with a custom Mock:

mock((+) => Mock(1)) do plus
    @assert 1 + 1 == 1
    @assert called_once_with(plus, 1, 1)
end

Mocking methods that match a given signature:

mock((+, Float64, Float64) => Mock((a, b) -> 2a + 2b)) do plus
    @assert 1 + 1 == 2
    @assert 2.0 + 2.0 == 8
    @assert called_once_with(plus, 2.0, 2.0)
end

Mocking with something other than a Mock:

mock((+) => (a, b) -> 2a + 2b) do _plus
    @assert 1 + 2 == 6
end

Using Filters

Oftentimes, you mock a function with a very specific idea of where you want that mocking to happen. It can be confusing when a call you didn't anticipate gets mocked somewhere deep in the call stack, botching everything. To avoid this, you can use filter functions like so:

f(x, y) = x + y
g(x, y) = f(x, y)
mock((+) => Mock((a, b) -> 2a + 2b); filters=[max_depth(2)]) do plus
    @assert f(1, 2) == 6  # The call depth of + here is 2.
    @assert g(3, 4) == 7  # Here, it's 3.
    @assert called_once_with(plus, 1, 2)
end

Filter functions take a single argument of type Metadata. If any filter rejects, then mocking is not performed. See Filter Functions for a list of included filters, as well as building blocks for you to create your own.

Reusing Contexts

Under the hood, this function creates a new Cassette Context on every call by default. This provides a nice clean mocking environment, but it can be slow to create and call new types and methods over and over. If you find yourself repeatedly mocking the same set of functions, you can specify a context name to reuse that context like so:

ctx = gensym()
mock(g -> @assert(!called(g)), ctx, get)
# This one is faster, especially when there's a lot going on in your mock blocks.
mock(g -> @assert(!called(g)), ctx, get)

Keyword Arguments

Mocking of functions with keyword arguments is fully supported when the following conditions are met:

Filter functions are not used

The context in use has no previously-mocked functions that are now unmocked

if a filter function rejects and calls a mocked function (instead of its mock), that call will have no keywords.

If you reuse a context that has previously mocked some function, unmocked calls to that function will have no keywords. For example:

kwfunc(; kwargs...) = nothing
calls_kwfunc(; kwargs...) = kwfunc(; kwargs...)
ctx = gensym()
mock(ctx, calls_kwfunc) do c
    calls_kwfunc(; x=1, y=2)
    @assert called_once_with(c; x=1, y=2)
end
mock(ctx, kwfunc) do k
    calls_kwfunc(; x=1, y=2)               # This will issue a warning.
    @assert called_once_with(k; x=1, y=2)  # This will fail!
end

In short, avoid using filters and reusing contexts when mocking functions that accept keywords.

Filter Functions

SimpleMock.max_depth — Function

max_depth(n::Int) -> Function

Create a filter that rejects when the current call depth is greater than n.

SimpleMock.min_depth — Function

min_depth(n::Int) -> Function

Create a filter that rejects when the current call depth is less than n.

SimpleMock.excluding — Function

excluding(args...) -> Function

Create a filter that rejects when the calling function or module is in args.

SimpleMock.including — Function

including(args...) -> Function

Create a filter that rejects when the calling function or module is not in args.

SimpleMock.Metadata — Type

Container for mocks and bookkeeping data. Instances of this type are aware of the call depth (current_depth) and the current function/module (current_function/current_module).

All filter functions take a single argument of this type.

SimpleMock.current_depth — Function

current_depth(::Metadata) -> Int

Return the current call depth (the size of the call stack). The depth is always positive, so the first function entered has a depth of 1.

SimpleMock.current_function — Function

current_function(::Metadata) -> Any

Return the current function (or other callable thing). In this case, "current" refers, somewhat counterintuitively, not to the function about to be called, but to the function that is about to call it.

To illustrate:

f(x) = x
g(x) = f(x)
x = g(1)

In this case, when the call to f is reached, the function that is calling f is g. Therefore, the "current" function is g.

SimpleMock.current_module — Function

current_module(m::Metadata) -> Module

Return the current module, where "current" has the same definition as in current_function.

Index

SimpleMock.SimpleMock

SimpleMock.Call

SimpleMock.Metadata

SimpleMock.Mock

SimpleMock.Predicate

SimpleMock.called

SimpleMock.called_last_with

SimpleMock.called_once

SimpleMock.called_once_with

SimpleMock.called_with

SimpleMock.calls

SimpleMock.current_depth

SimpleMock.current_function

SimpleMock.current_module

SimpleMock.excluding

SimpleMock.has_call

SimpleMock.has_calls

SimpleMock.including

SimpleMock.last_call

SimpleMock.max_depth

SimpleMock.min_depth

SimpleMock.mock

SimpleMock.ncalls

SimpleMock.reset!